Adds manualStitchOrder and makes some changes to stitched image file formats
Adds new "manualStitchOrder" flag to StitchFullEmbryoImages to allow user input on choice of tile stitching order. This was necessary because the mid-saggital images often result in broken stitching when overlap between tiles is all in the middle of the yolk.

+++ src/axisRegistry/APTileScan/NewTileArrayFromMetadata.m
@@ -58,8 +58,8 @@ function tile_array = NewTileArrayFromMetadata(Prefix, ID)
         ImageSlices = generateHisSlicesTileScan(LIFImages, NSlices(n),...
             NChannels, fiducialChannel, framesIndex, n);
         temp = max(ImageSlices, [], 3);
-        zstacks{n} = ImageSlices;
-        tiles{n} = temp;
+        zstacks{n} = uint16(ImageSlices);
+        tiles{n} = uint16(temp);
     end
